use one beaker session within tera-model-edit
