+++ packages/client/hmi-client/vite.config.local.ts
@@ -14,13 +14,13 @@ config.server.proxy = {
 		changeOrigin: true
 	},
 	'^/beaker_ws': {
-		target: `ws://${localhost}:3050`,
+		target: `ws://${localhost}:8888`,
 		rewrite: (path_str) => path_str.replace(/^\/beaker_ws/, ''),
 		changeOrigin: true,
 		ws: true
 	},
 	'^/beaker': {
-		target: `http://${localhost}:3050`,
+		target: `http://${localhost}:8888`,
 		rewrite: (path_str) => path_str.replace(/^\/beaker/, ''),
 		changeOrigin: true
 	}
